finished scooping interface
